jules_v2_optimized/requirements.txt
google-generativeai>=0.3.0
tenacity>=8.2.0
aiofiles>=23.2.1